Fix tests (so to speak)
