+++ packages/esm-service-queues-app/src/metrics/metrics.resource.ts
@@ -56,3 +56,24 @@ export function useAverageWaitTime(serviceUuid: string, statusUuid: string) {
     mutate,
   };
 }
+
+export function useServiceMetricsCount(service: string, location: string) {
+  const status = 'Waiting';
+  const apiUrl =
+    `${restBaseUrl}/queue-entry-metrics?status=${status}&isEnded=false` +
+    (service ? `&service=${service}` : '') +
+    (location ? `&location=${location}` : '');
+
+  const { data } = useSWR<
+    {
+      data: {
+        count: number;
+      };
+    },
+    Error
+  >(service ? apiUrl : null, openmrsFetch);
+
+  return {
+    serviceCount: data ? data?.data?.count : 0,
+  };
+}
